scripts/pyproject.toml
[project]
name = "manga-metadata"
version = "0.1.0"
description = "Tools for processing manga PDFs with Gemini and extracting per-page metadata."
requires-python = ">=3.10"

dependencies = [
    "pymupdf",
    "google-generativeai",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["process_page", "process_chapter"]

[tool.uv]
package = true


